backends established

+++ controllers/apiRoutes/opinionRoutes.js
@@ -25,8 +25,8 @@ router.get("/:id", (req, res) => {
 })
 
 router.post("/", (req, res) => {
-    if(!req.session.user){
-        return res.status(403).json({err:"Login is required."})
+    if (!req.session.user) {
+        return res.status(403).json({ err: "Login is required." })
     }
     Opinion.create({
         like: req.body.like,
@@ -44,31 +44,48 @@ router.post("/", (req, res) => {
 })
 
 router.put("/:id", (req, res) => {
-    Opinion.update({
-        like: req.body.like,
-        review: req.body.review,
-        //user_id: req.body.user_id,
-        //maybe user.id?????
-        //opinion_id: req.body.opinion_id
-        //maybe post.id?????
-    }, {
-        where: {
-            id: req.params.id
+    if (!req.session.user) {
+        return res.status(403).json({ err: "Login is required." })
+    }
+    Opinion.findByPk(req.params.id).then(foundOpinion => {
+        if (req.session.user.id !== foundOpinion.UserId) {
+            return res.status(403).json({ err: "This review cannot be edited." })
         }
-    }).then(updatedOpinion => {
-        if (updatedOpinion[0]) {
-            res.json(updatedOpinion)
-        } else {
-            res.status(404).json({ err: "There were no posts found!" })
-        }
-        res.json(updatedOpinion)
-    }).catch(err => {
+
+        Opinion.update({
+            like: req.body.like,
+            review: req.body.review,
+            //maybe user.id?????
+            opinion_id: req.body.opinion_id
+            //maybe post.id?????
+        },
+            {
+                where: {
+                    id: req.params.id
+                }
+            }
+        )
+            .then(updatedOpinion => {
+                if (updatedOpinion[0]) {
+                    res.json(updatedOpinion)
+                } else {
+                    res.status(404).json({ err: "There were no posts found!" })
+                }
+                res.json(updatedOpinion)
+            }).catch(err => {
+                console.error(err);
+                res.status(500).json({ err });
+            });
+    }).catch(err=> {
         console.error(err);
         res.status(500).json({ err });
-    })
-})
+    });
+});
 
 router.delete("/:id", (req, res) => {
+    if (!req.session.user) {
+        return res.status(403).json({ err: "Login is required." })
+    }
     Opinion.destroy({
         where: {
             id: req.params.id
